+++ django_cron/management/commands/runcrons.py
@@ -1,10 +1,11 @@
-import sys
 from optparse import make_option
 import traceback
+from datetime import timedelta
 
 from django.core.management.base import BaseCommand
 from django.conf import settings
-from django_cron import CronJobManager, get_class
+from django_cron import CronJobManager, get_class, get_current_time
+from django_cron.models import CronJobLog
 try:
     from django.db import close_old_connections as close_connection
 except ImportError:
@@ -12,7 +13,6 @@ except ImportError:
 
 
 DEFAULT_LOCK_TIME = 24 * 60 * 60  # 24 hours
-
 
 
 class Command(BaseCommand):
@@ -32,15 +32,20 @@ class Command(BaseCommand):
             cron_class_names = getattr(settings, 'CRON_CLASSES', [])
 
         try:
-            crons_to_run = map(lambda x: get_class(x), cron_class_names)
-        except:
+            crons_to_run = [get_class(x) for x in cron_class_names]
+        except Exception:
             error = traceback.format_exc()
-            print('Make sure these are valid cron class names: %s\n%s' % (cron_class_names, error))
-            sys.exit()
+            self.stdout.write('Make sure these are valid cron class names: %s\n%s' % (cron_class_names, error))
+            return
 
         for cron_class in crons_to_run:
-            run_cron_with_cache_check(cron_class, force=options['force'],
-                silent=options['silent'])
+            run_cron_with_cache_check(
+                cron_class,
+                force=options['force'],
+                silent=options['silent']
+            )
+
+        clear_old_log_entries()
         close_connection()
 
 
@@ -56,3 +61,11 @@ def run_cron_with_cache_check(cron_class, force=False, silent=False):
     with CronJobManager(cron_class, silent) as manager:
         manager.run(force)
 
+
+def clear_old_log_entries():
+    """
+    Removes older log entries, if the appropriate setting has been set
+    """
+    if hasattr(settings, 'DJANGO_CRON_DELETE_LOGS_OLDER_THAN'):
+        delta = timedelta(days=settings.DJANGO_CRON_DELETE_LOGS_OLDER_THAN)
+        CronJobLog.objects.filter(end_time__lt=get_current_time() - delta).delete()