+++ django_cron/tests.py
@@ -1,9 +1,9 @@
 import threading
 from time import sleep
+from datetime import timedelta
 
 from django import db
 from django.utils import unittest
-from django_cron.models import CronJobLog
 from django.core.management import call_command
 from django.test.utils import override_settings
 from django.test.client import Client
@@ -12,6 +12,26 @@ from django.contrib.auth.models import User
 
 from freezegun import freeze_time
 
+from django_cron.helpers import humanize_duration
+from django_cron.models import CronJobLog
+
+
+class OutBuffer(object):
+    content = []
+    modified = False
+    _str_cache = ''
+
+    def write(self, *args):
+        self.content.extend(args)
+        self.modified = True
+
+    def str_content(self):
+        if self.modified:
+            self._str_cache = ''.join((str(x) for x in self.content))
+            self.modified = False
+
+        return self._str_cache
+
 
 class TestCase(unittest.TestCase):
 
@@ -19,7 +39,8 @@ class TestCase(unittest.TestCase):
     error_cron = 'test_crons.TestErrorCronJob'
     five_mins_cron = 'test_crons.Test5minsCronJob'
     run_at_times_cron = 'test_crons.TestRunAtTimesCronJob'
-    wait_3sec_cron = 'test_crons.Wiat3secCronJob'
+    wait_3sec_cron = 'test_crons.Wait3secCronJob'
+    does_not_exist_cron = 'ThisCronObviouslyDoesntExist'
     test_failed_runs_notification_cron = 'django_cron.cron.FailedRunsNotificationCronJob'
 
     def setUp(self):
@@ -34,6 +55,15 @@ class TestCase(unittest.TestCase):
         logs_count = CronJobLog.objects.all().count()
         call_command('runcrons', self.error_cron, force=True)
         self.assertEqual(CronJobLog.objects.all().count(), logs_count + 1)
+
+    def test_not_exists_cron(self):
+        logs_count = CronJobLog.objects.all().count()
+        out_buffer = OutBuffer()
+        call_command('runcrons', self.does_not_exist_cron, force=True, stdout=out_buffer)
+
+        self.assertIn('Make sure these are valid cron class names', out_buffer.str_content())
+        self.assertIn(self.does_not_exist_cron, out_buffer.str_content())
+        self.assertEqual(CronJobLog.objects.all().count(), logs_count)
 
     @override_settings(DJANGO_CRON_LOCK_BACKEND='django_cron.backends.lock.file.FileLock')
     def test_file_locking_backend(self):
@@ -58,7 +88,6 @@ class TestCase(unittest.TestCase):
 
     def test_runs_at_time(self):
         logs_count = CronJobLog.objects.all().count()
-
         with freeze_time("2014-01-01 00:00:01"):
             call_command('runcrons', self.run_at_times_cron)
         self.assertEqual(CronJobLog.objects.all().count(), logs_count + 1)
@@ -73,13 +102,13 @@ class TestCase(unittest.TestCase):
 
     def test_admin(self):
         password = 'test'
-        user = User.objects.create_superuser('test', 'test@tivix.com',
-            password)
+        user = User.objects.create_superuser(
+            'test',
+            'test@tivix.com',
+            password
+        )
         self.client = Client()
         self.client.login(username=user.username, password=password)
-
-        # get list of CronJobLogs
-        url = reverse('admin:django_cron_cronjoblog_changelist')
 
         # edit CronJobLog object
         call_command('runcrons', self.success_cron, force=True)
@@ -136,3 +165,17 @@ class TestCase(unittest.TestCase):
         call_command('runcrons', self.test_failed_runs_notification_cron)
 
         self.assertEqual(CronJobLog.objects.all().count(), logs_count + 11)
+
+    def test_humanize_duration(self):
+        test_subjects = (
+            (timedelta(days=1, hours=1, minutes=1, seconds=1), '1 day, 1 hour, 1 minute, 1 second'),
+            (timedelta(days=2), '2 days'),
+            (timedelta(days=15, minutes=4), '15 days, 4 minutes'),
+            (timedelta(), '< 1 second'),
+        )
+
+        for duration, humanized in test_subjects:
+            self.assertEqual(
+                humanize_duration(duration),
+                humanized
+            )